Deletes interface files
