+++ src/main/java/edu/hawaii/its/groupings/access/AuthorizationService.java
@@ -1,6 +1,65 @@
 package edu.hawaii.its.groupings.access;
 
-public interface AuthorizationService {
-    RoleHolder fetchRoles(String uhUuid, String uid);
+import org.apache.commons.logging.Log;
+import org.apache.commons.logging.LogFactory;
+import org.jasig.cas.client.authentication.SimplePrincipal;
+import edu.hawaii.its.api.controller.GroupingsRestController;
+import edu.hawaii.its.groupings.service.UhUuidCheckerService;
 
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.PropertySource;
+import org.springframework.http.ResponseEntity;
+import org.springframework.stereotype.Service;
+
+import java.security.Principal;
+
+@Service
+@PropertySource("classpath:custom.properties")
+public class AuthorizationService {
+
+    @Autowired
+    GroupingsRestController groupingsRestController;
+
+    @Autowired
+    UhUuidCheckerService uhUuidCheckerService;
+    private static final Log logger = LogFactory.getLog(AuthorizationService.class);
+
+    /**
+     * Assign roles to user
+     */
+
+    public RoleHolder fetchRoles(String uhUuid, String uid) {
+        RoleHolder roleHolder = new RoleHolder();
+        Principal principal = new SimplePrincipal(uhUuid);
+        roleHolder.add(Role.ANONYMOUS);
+
+        //Determine if uhUuid is valid
+        if (uhUuidCheckerService.isValidUhUuid(uhUuid, uid)) {
+            roleHolder.add(Role.UH);
+        }
+
+        //Determine if user is an owner.
+        if (checkResult(groupingsRestController.hasOwnerPrivs(principal))) {
+            roleHolder.add(Role.OWNER);
+        }
+
+        //Determine if a user is an admin.
+        if (checkResult(groupingsRestController.hasAdminPrivs(principal))) {
+            roleHolder.add(Role.ADMIN);
+        }
+
+        logger.info("fetchRoles: username: " + uid + " " + roleHolder.getAuthorities() + ";");
+        return roleHolder;
+    }
+
+    /**
+     * Return a boolean based on parsed response.
+     */
+    private boolean checkResult(ResponseEntity response) {
+        if (response == null || response.getBody() == null) {
+            return false;
+        }
+
+        return Boolean.parseBoolean((String) response.getBody());
+    }
 }
