+++ src/main/java/edu/hawaii/its/groupings/access/UhAttributes.java
@@ -1,27 +1,127 @@
 package edu.hawaii.its.groupings.access;
 
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Collections;
+import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-public interface UhAttributes {
+public class UhAttributes  {
 
-    // Methods specific to UH results.
+    private Map<String, List<String>> uhAttributeMap = new HashMap<>();
+    private final String uid; // CAS login username.
+    private final Map<?, ?> map; // Original CAS results.
 
-    String getUid();
+    // Constructor.
+    public UhAttributes() {
+        this(new HashMap<>());
+    }
 
-    String getUhUuid();
+    // Constructor.
+    public UhAttributes(Map<?, ?> map) {
+        this("", map);
+    }
 
-    String getName();
+    // Constructor.
+    public UhAttributes(String uid, Map<?, ?> map) {
+        this.uid = uid != null ? uid : "";
+        this.map = map;
+        if (map != null) {
+            for (Object key : map.keySet()) {
+                if (key instanceof String) {
+                    String k = ((String) key).toLowerCase();
+                    Object v = map.get(key);
+                    if (v != null) {
+                        if (v instanceof String) {
+                            uhAttributeMap.put(k, Arrays.asList((String) v));
+                        } else if (v instanceof List) {
+                            List<String> lst = new ArrayList<>();
+                            for (Object o : (List<?>) v) {
+                                if (o != null && o instanceof String) {
+                                    lst.add((String) o);
+                                }
+                            }
+                            uhAttributeMap.put(k, lst);
+                        }
+                    }
+                }
+            }
+        }
+    }
 
-    List<String> getMail();
+    public String getName() {
+        return getValue("cn");
+    }
 
-    List<String> getAffiliation();
+    public String getUsername() {
+        //return username;
+        return "";
+    }
 
-    // Generic methods.
+    public String getUid() {
+        List<String> values = uhAttributeMap.get("uid");
+        if (values != null) {
+            // Check expected case first.
+            if (values.size() == 1) {
+                return values.get(0); // We are done.
+            }
 
-    String getValue(String name);
+            if (values.size() > 1) {
+                // More than one uid in the results.
+                // Try to match up with the username.
+                for (String s : values) {
+                    if (s.equals(getUsername())) {
+                        return s;
+                    }
+                }
 
-    List<String> getValues(String name);
+                // Couldn't match up username with uid,
+                // so just return first value.
+                return values.get(0); // We are done.
+            }
+        }
 
-    Map<?, ?> getMap();
+        return ""; // Didn't find anything.
+    }
+
+    public String getUhUuid() {
+        return getValue("uhUuid");
+    }
+
+    public List<String> getMail() {
+        return getValues("mail");
+    }
+
+    public List<String> getAffiliation() {
+        return getValues("eduPersonAffiliation");
+    }
+
+    public List<String> getValues(String name) {
+        List<String> results = uhAttributeMap.get(toLowerCase(name));
+        if (results != null) {
+            return Collections.unmodifiableList(results);
+        }
+        return Collections.emptyList();
+    }
+
+    public String getValue(String name) {
+        List<String> results = getValues(name);
+        return results.isEmpty() ? "" : results.get(0);
+    }
+
+    public Map<?, ?> getMap() {
+        return Collections.unmodifiableMap(map);
+    }
+
+    private String toLowerCase(String s) {
+        return (s != null) ? s.toLowerCase() : s;
+    }
+
+    public String toString() {
+        return "UhCasAttributes [username=" + uid
+                + ", uhAttributeMap=" + uhAttributeMap
+                + ", map=" + map + "]";
+    }
+
 }
