+++ src/main/java/edu/hawaii/its/groupings/access/UserContextService.java
@@ -1,11 +1,44 @@
 package edu.hawaii.its.groupings.access;
 
-public interface UserContextService {
-    User getCurrentUser();
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContext;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.stereotype.Service;
 
-    String getCurrentUid();
+@Service
+public class UserContextService {
 
-    String getCurrentUhUuid();
+    public User getCurrentUser() {
+        SecurityContext securityContext = SecurityContextHolder.getContext();
+        Authentication authentication = securityContext.getAuthentication();
+        if (authentication != null) {
+            Object principal = authentication.getPrincipal();
+            if (principal instanceof UserDetails) {
+                return (User) authentication.getPrincipal();
+            }
+        }
+        return new AnonymousUser();
+    }
 
-    void setCurrentUhUuid(String uhUuid);
+    public String getCurrentUid() {
+        User user = getCurrentUser();
+        return user != null ? user.getUid() : "";
+    }
+
+    public String getCurrentUhUuid() {
+        User user = getCurrentUser();
+        return user != null ? user.getUhUuid() : "";
+    }
+
+    public void setCurrentUhUuid(String uhUuid) {
+        User user = getCurrentUser();
+        if (null != user && user.hasRole(Role.ADMIN)) {
+            user.setUhUuid(uhUuid);
+        }
+    }
+
+    public String toString() {
+        return "UserContextServiceImpl [context=" + SecurityContextHolder.getContext() + "]";
+    }
 }
