+++ src/test/java/edu/hawaii/its/api/controller/GroupingsRestControllerTest.java
@@ -33,6 +33,7 @@ import java.util.List;
 import java.util.Map;
 
 
+import edu.hawaii.its.api.service.HttpRequestService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -48,7 +49,6 @@ import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.web.context.WebApplicationContext;
 
-import edu.hawaii.its.api.service.HttpRequestService;
 import edu.hawaii.its.groupings.configuration.Realm;
 import edu.hawaii.its.groupings.configuration.SpringBootWebApplication;
 import edu.hawaii.its.groupings.controller.WithMockUhUser;
